Start working on a datepicker web component
Web components seem to be the new hot thing, so let's try them out. This
patch adds the beginnings of a datepicker element, which displays either
the current month or a month specified by optional 'month' and 'year'
attributes in HTML. The month can be dynamically changed via buttons at
the top of the element.